Css.Selector.Core

Description

A module that defines the tree of types to represent and manipulate a css selector. These data types are members of several typeclasses to make these more useful.

ToCssSelector typeclass

class ToCssSelector a where

A class that defines that the given type can be converted to a css selector value, and has a certain specificity.

Minimal complete definition

toCssSelector, toSelectorGroup, specificity', toPattern

Methods

toCssSelector

Arguments

:: a
The given object for which we calculate the css selector.

-> Text
The css selector text for the given object.

Convert the given element to a Text object that contains the css selector.

toSelectorGroup

Arguments

:: a
The item to lift to a SelectorGroup

-> SelectorGroup
The value of a SelectorGroup of which the object is the selective part.

Lift the given ToCssSelector type object to a SelectorGroup, which is the "root type" of the css selector hierarchy.

specificity'

Arguments

:: a
The item for which we calculate the specificity level.

-> SelectorSpecificity
The specificity level of the given item. Convert the given ToCssSelector item to a Pat pattern, such that we can use it in functions.

Calculate the specificity of the css selector by returing a SelectorSpecificity object.

toPattern

Arguments

:: a
The item to convert to a Pat.

-> Pat
The pattern that is generated that will match only items equal to the given object. Convert the given ToCssSelector item to an item in a more normalized form. A normalization is idempotent: applying this multiple times will have the same effect as applying it once.

normalize

Arguments

:: a
The item to normalize.

-> a
A normalized variant of the given item. This will filter the same objects, and have the same specificity.

Selectors and combinators

data Selector

The type of a single selector. This is a sequence of SelectorSequences that are combined with a SelectorCombinator.

Constructors

Selector SelectorSequence
Convert a given SelectorSequence to a Selector.

Combined SelectorSequence SelectorCombinator Selector
Create a combined selector where we have a SelectorSequence that is combined with a given SelectorCombinator to a Selector.

data SelectorCombinator

A type that contains the possible ways to combine SelectorSequences.

Constructors

Descendant
The second tag is a descendant of the first one, denoted in css with a space.

Child
The second tag is the (direct) child of the first one, denoted with a > in css.

DirectlyPreceded
The second tag is directly preceded by the first one, denoted with a + in css.

Preceded
The second tag is preceded by the first one, denoted with a ~ in css.

newtype SelectorGroup

The root type of a css selector. This is a comma-separated list of selectors.

Constructors

SelectorGroup

Fields

unSelectorGroup :: NonEmpty Selector

Unwrap the given NonEmpty list of Selectors from the SelectorGroup object.

data SelectorSequence

A SelectorSequence is a TypeSelector (that can be Universal) followed by zero, one or more SelectorFilters these filter the selector further, for example with a Hash, a Class, or an Attrib.

Constructors

SimpleSelector TypeSelector
Convert a TypeSelector into a SimpleSelector.

Filter SelectorSequence SelectorFilter
Apply an additional SelectorFilter to the SelectorSequence.

combinatorText

Arguments

:: SelectorCombinator
The given SelectorCombinator to retrieve the css token for.

-> Text
The css selector token that is used for the given SelectorCombinator.

Convert the SelectorCombinator to the equivalent css selector text. A space for Descendant, a > for Child, a + for DirectlyPreceded, and a ~ for Preceded

combine

Arguments

:: SelectorCombinator
The SelectorCombinator that is applied between the two Selectors.

-> Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A Selector that is a combination of the left Selector and the right Selector with the given SelectorCombinator.

Combines two Selectors with the given SelectorCombinator.

(.>)

Arguments

:: Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A selector that is the combination of the left Selector and the right Selector through Child.

Combines two Selectors with the Child combinator.

(.+)

Arguments

:: Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A selector that is the combination of the left Selector and the right Selector through DirectlyPreceded.

Combines two Selectors with the DirectlyPreceded combinator.

(.~)

Arguments

:: Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A selector that is the combination of the left Selector and the right Selector through Preceded.

Combines two Selectors with the Preceded combinator.

Filters

data SelectorFilter

A type that sums up the different ways to filter a type selector: with an id (hash), a class, and an attribute.

Constructors

SHash Hash
A Hash object as filter.

SClass Class
A Class object as filter.

SAttrib Attrib
An Attrib object as filter.

filters

Arguments

:: SelectorSequence
The given SelectorSequence to analyze.

-> [SelectorFilter]
The given list of SelectorFilters applied, this can be empty.

Obtain the list of filters that are applied in the given SelectorSequence.

filters'

Arguments

:: SelectorSequence
The given SelectorSequence to analyze.

-> [SelectorFilter]
The given list of SelectorFilters applied in reversed order, this can be empty.

Obtain the list of filters that are applied in the given SelectorSequence in reversed order.

addFilters

Arguments

:: SelectorSequence
The SelectorSequence to apply the filter on.

-> [SelectorFilter]
The list of SelectorFilters to apply on the SelectorSequence.

-> SelectorSequence
A modified SelectorSequence where we applied the list of SelectorFilters.

Add a given list of SelectorFilters to the given SelectorSequence. The filters are applied left-to-right.

(.:)

Arguments

:: SelectorSequence
The SelectorSequence to apply the filter on.

-> [SelectorFilter]
The list of SelectorFilters to apply on the SelectorSequence.

-> SelectorSequence
A modified SelectorSequence where we applied the list of SelectorFilters.

An infix variant of the addFilters function.

Namespaces

data Namespace

The namespace of a css selector tag. The namespace can be NAny (all possible namespaces), or a namespace with a given text (this text can be empty).

Constructors

NAny
A typeselector part that specifies that we accept all namespaces, in css denoted with *.

Namespace Text
A typselector part that specifies that we accept a certain namespace name.

pattern NEmpty :: Namespace

The empty namespace. This is not the wildcard namespace (*). This is a bidirectional namespace and can thus be used in expressions as well.

Type selectors

data ElementName

The element name of a css selector tag. The element name can be EAny (all possible tag names), or an element name with a given text.

Constructors

EAny
A typeselector part that specifies that we accept all element names, in css denoted with *.

ElementName Text
A typeselector part that specifies that we accept a certain element name.

data TypeSelector

A typeselector is a combination of a selector for a namespace, and a selector for an element name. One, or both can be a wildcard.

Constructors

TypeSelector

Fields

selectorNamespace :: Namespace

The selector for the namespace.

elementName :: ElementName

The selector for the element name.

pattern Universal :: TypeSelector

The universal type selector: a selector that matches all types in all namespaces (including the empty namespace). This pattern is bidirectional and thus can be used in expressions as well.

(.|)

Arguments

:: Namespace
The Namespace for the TypeSelector.

-> ElementName
The ElementName for the TypeSelector.

-> TypeSelector
A TypeSelector object constructed with the Namespace and ElementName.

Construct a TypeSelector with a given Namespace and ElementName.

Attributes

data Attrib

A css attribute can come in two flavors: either a constraint that the attribute should exists, or a constraint that a certain attribute should have a certain value (prefix, suffix, etc.).

Constructors

Exist AttributeName
A constraint that the given AttributeName should exist.

Attrib AttributeName AttributeCombinator AttributeValue
A constraint about the value associated with the given AttributeName.

data AttributeCombinator

The possible ways to match an attribute with a given value in a css selector.

Constructors

Exact
The attribute has exactly the value of the value, denoted with = in css.

Include
The attribute has a whitespace separated list of items, one of these items is the value, denoted with ~= in css.

DashMatch
The attribute has a hyphen separated list of items, the first item is the value, denoted with |= in css.

PrefixMatch
The value is a prefix of the value in the attribute, denoted with ^= in css.

SuffixMatch
The value is a suffix of the value in the attribute, denoted with $= in css.

SubstringMatch
The value is a substring of the value in the attribute, denoted with *= in css.

data AttributeName

An attribute name is a name that optionally has a namespace, and the name of the attribute.

Constructors

AttributeName

Fields

attributeNamespace :: Namespace

The namespace to which the attribute name belongs. This can be NAny as well.

attributeName :: Text

The name of the attribute over which we make a claim.

type AttributeValue = Text

We use Text as the type to store an attribute value.

(.=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted to be exactly the given value.

(.~=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute is a whitespace seperated list of items, and the value is one of these items.

(.|=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute is a dash seperated list of items, and the value is the first of these items.

(.^=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute has as prefix the given AttributeValue.

(.$=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute has as suffix the given AttributeValue.

(.*=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute has as substring the given AttributeValue.

attrib

Arguments

:: AttributeCombinator
The AttributeCombinator that specifies the required relation between the attribute and a value.

-> AttributeName
The name of an attribute to filter.

-> AttributeValue
The value of the attribute to filter.

-> Attrib
The result is an Attrib object that will filter the given AttributeName with the given AttributeCombinator.

A flipped version of the Attrib data constructor, where one first specifies the conbinator, then the AttributeName and finally the value.

attributeCombinatorText

Arguments

:: AttributeCombinator
The AttributeCombinator for which we obtain the corresponding css selector text.

-> AttributeValue
The css selector text for the given AttributeCombinator.

Convert the given AttributeCombinator to its css selector counterpart.

Classes

newtype Class

A css class, this is wrapped in a data type. The type only wraps the class name, not the dot prefix.

Constructors

Class

Fields

unClass :: Text

Obtain the name from the class.

(...)

Arguments

:: SelectorSequence
The given 'SelectorSequence to filter.

-> Class
The given Class to filter the SelectorSequence further.

-> SelectorSequence
A SelectorSequence that is filtered additionally with the given Class.

Filter a given SelectorSequence with a given Class.

Hashes

newtype Hash

A css hash (used to match an element with a given id). The type only wraps the hash name, not the hash (#) prefix.

Constructors

Hash

Fields

unHash :: Text

Obtain the name from the hash.

(.#)

Arguments

:: SelectorSequence
The given SelectorSequence to filter.

-> Hash
The given Hash to filter the SelectorSequence further.

-> SelectorSequence
A SelectorSequence that is filtered additionally with the given Hash.

Filter a given SelectorSequence with a given Hash.

Specificity

data SelectorSpecificity

A datastructure that specifies the selectivity of a css selector. The specificity is calculated based on three integers: a, b and c.

The specificity is calculated with 100*a+10*b+c where a, b and c count certain elements of the css selector.

Constructors

SelectorSpecificity Int Int Int
Create a SelectorSpecificity object with a given value for a, b, and c.

specificity

Arguments

:: ToCssSelector a

=> a
The object for which we evaluate the specificity.

-> Int
The specificity level as an Int value.

Calculate the specificity of a ToCssSelector type object. This is done by calculating the SelectorSpecificity object, and then calculating the value of that object.

specificityValue

Arguments

:: SelectorSpecificity
The SelectorSpecificity to calculate the specificity value from.

-> Int
The specificity level of the SelectorSpecificity. If the value is higher, the rules in the css selector take precedence.

Calculate the specificity value of the SelectorSpecificity