Css3.Selector.Core

Description

A module that defines the tree of types to represent and manipulate a css selector. These data types are members of several typeclasses to make these more useful.

ToCssSelector typeclass

class ToCssSelector a where

A class that defines that the given type can be converted to a css selector value, and has a certain specificity.

Minimal complete definition

toCssSelector, toSelectorGroup, specificity', toPattern

Methods

toCssSelector

Arguments

:: a
The given object for which we calculate the css selector.

-> Text
The css selector text for the given object.

Convert the given element to a Text object that contains the css selector.

toSelectorGroup

Arguments

:: a
The item to lift to a SelectorGroup

-> SelectorGroup
The value of a SelectorGroup of which the object is the selective part.

Lift the given ToCssSelector type object to a SelectorGroup, which is the "root type" of the css selector hierarchy.

specificity'

Arguments

:: a
The item for which we calculate the specificity level.

-> SelectorSpecificity
The specificity level of the given item. Convert the given ToCssSelector item to a Pat pattern, such that we can use it in functions.

Calculate the specificity of the css selector by returing a SelectorSpecificity object.

toPattern

Arguments

:: a
The item to convert to a Pat.

-> Pat
The pattern that is generated that will match only items equal to the given object. Convert the given ToCssSelector item to an item in a more normalized form. A normalization is idempotent: applying this multiple times will have the same effect as applying it once.

normalize

Arguments

:: a
The item to normalize.

-> a
A normalized variant of the given item. This will filter the same objects, and have the same specificity.

Selectors and combinators

data Selector

The type of a single selector. This is a sequence of SelectorSequences that are combined with a SelectorCombinator.

Constructors

Selector PseudoSelectorSequence
Convert a given SelectorSequence to a Selector.

Combined PseudoSelectorSequence SelectorCombinator Selector
Create a combined selector where we have a SelectorSequence that is combined with a given SelectorCombinator to a Selector.

data SelectorCombinator

A type that contains the possible ways to combine SelectorSequences.

Constructors

Descendant
The second tag is a descendant of the first one, denoted in css with a space.

Child
The second tag is the (direct) child of the first one, denoted with a > in css.

DirectlyPreceded
The second tag is directly preceded by the first one, denoted with a + in css.

Preceded
The second tag is preceded by the first one, denoted with a ~ in css.

newtype SelectorGroup

The root type of a css selector. This is a comma-separated list of selectors.

Constructors

SelectorGroup

Fields

unSelectorGroup :: NonEmpty Selector

Unwrap the given NonEmpty list of Selectors from the SelectorGroup object.

data PseudoElement

An enum type that contains the possible pseudo elements. A pseudo element is specified by two colon characters (::), followed by the name of the pseudo element. The After, Before, FirstLine and FirstLetter can be written with a single colon for backwards compatibility with CSS 1 and CSS 2.

Constructors

After
The ::after pseudo-elements can be used to describe generated content after an element’s content.

Before
The ::before pseudo-element can be used to describe generated content before an element’s content.

FirstLetter
The ::first-line pseudo-element describes the contents of the first formatted line of an element.

FirstLine
The ::first-letter pseudo-element represents the first letter of an element, if it is not preceded by any other content (such as images or inline tables) on its line.

Marker
The ::marker pseudo-element selects the markers of list items.

Placeholder
The ::placeholder pseudo-element selects form elements with placeholder text, and let you style the placeholder text.

Selection
The ::selection pseudo-element matches the portion of an element that is selected by a user.

data PseudoSelectorSequence

A SelectorSequence with an optional PseudoElement at the end. Each element of a Selector can have at most one PseudoElement.

Constructors

Sequence SelectorSequence
A data constructor where there is no optional PseudoElement involved.

SelectorSequence :.:: PseudoElement
A data constructor for a SelectorSequence with a PseudoElement.

(.::)

Arguments

:: SelectorSequence
The given SelectorSequence to which we add the pseudo element.

-> PseudoElement
The given PseudoElement to add to the SelectorSequence.

-> PseudoSelectorSequence
The corresponding PseudoSelectorSequence.

Add a given PseudoElement to the given SelectorSequence to produce a PseudoSelectorSequence. Since a PseudoElement is an instance of IsString, this can thus be used to combine string literals.

data PseudoClass

A data type that contains the possible pseudo classes. In a CSS selector the pseudo classes are specified with a single colon, for example :active. These filter on the state of the items. A full list of pseudo classes is available here.

Constructors

Active
The :active pseudo class.

Checked
The :checked pseudo class.

Default
The :default pseudo class.

Disabled
The :disabled pseudo class.

Empty
The :empty pseudo class.

Enabled
The :enabled pseudo class.

Focus
The :focus pseudo class.

Fullscreen
The :fullscreen pseudo class.

Hover
The :hover pseudo class.

Indeterminate
The :indeterminate pseudo class.

InRange
The :in-range pseudo class.

Invalid
The :invalid pseudo class.

Lang Language
The :lang(…) pseudo class, the language parameter is at the moment a Text object, but only uppercase, lowercase and hyphens are characters that can be parsed.

Link
The :link pseudo class.

NthChild Nth
The :nth-child(…) pseudo class, if the Nth parameter is One, then it is equivalent to :first-child.

NthLastChild Nth
The :nth-last-child(…) pseudo class, if the Nth parameter is One, then it is equivalent to :last-child.

NthLastOfType Nth
The :nth-last-of-type(…) pseudo class, if the Nth parameter is One, then it is equivalent to :last-of-type.

NthOfType Nth
The :nth-of-type(…) pseudo class, if the Nth parameter is One, then it is equivalent to :first-of-type.

OnlyOfType
The :only-of-type pseudo class.

OnlyChild
The :only-child pseudo class.

Optional
The :optional pseudo class.

OutOfRange
The :out-of-range pseudo class.

ReadOnly
The :read-only pseudo class.

ReadWrite
The :rad-write pseudo class.

Required
The :required pseudo class.

Root
The :root pseudo class.

Target
The :target pseudo class.

Valid
The :valid pseudo class.

Visited
The :visited pseudo class.

(.:)

Arguments

:: SelectorSequence
The given SelectorSequence to filter.

-> PseudoClass
The given PseudoClass to filter the SelectorSequence further.

-> SelectorSequence
A SelectorSequence that is filtered additionally with the given PseudoClass.

Filter a given SelectorSequence with a given PseudoClass.

pattern FirstChild :: PseudoClass

A pattern synonym for :nth-child(1). If NthChild (Nth 0 1) is used, then this will render as :first-child.

pattern FirstOfType :: PseudoClass

A pattern synonym for :nth-of-type(1). If NthOfType (Nth 0 1) is used, then this will render as :first-of-type.

pattern LastChild :: PseudoClass

A pattern synonym for :nth-last-child(1). If NthLastChild (Nth 0 1) is used, then this will render as :last-child.

pattern LastOfType :: PseudoClass

A pattern synonym for :nth-last-of-type(1). If NthLastOfType (Nth 0 1) is used, then this will render as :last-of-type.

type Language = Text

We use Text to specify the language in the :lang(…) pseudo class.

data SelectorSequence

A SelectorSequence is a TypeSelector (that can be Universal) followed by zero, one or more SelectorFilters these filter the selector further, for example with a Hash, a Class, or an Attrib.

Constructors

SimpleSelector TypeSelector
Convert a TypeSelector into a SimpleSelector.

Filter SelectorSequence SelectorFilter
Apply an additional SelectorFilter to the SelectorSequence.

combinatorText

Arguments

:: SelectorCombinator
The given SelectorCombinator to retrieve the css token for.

-> Text
The css selector token that is used for the given SelectorCombinator.

Convert the SelectorCombinator to the equivalent css selector text. A space for Descendant, a > for Child, a + for DirectlyPreceded, and a ~ for Preceded

combine

Arguments

:: SelectorCombinator
The SelectorCombinator that is applied between the two Selectors.

-> Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A Selector that is a combination of the left Selector and the right Selector with the given SelectorCombinator.

Combines two Selectors with the given SelectorCombinator.

(.>)

Arguments

:: Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A selector that is the combination of the left Selector and the right Selector through Child.

Combines two Selectors with the Child combinator.

(.+)

Arguments

:: Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A selector that is the combination of the left Selector and the right Selector through DirectlyPreceded.

Combines two Selectors with the DirectlyPreceded combinator.

(.~)

Arguments

:: Selector
The left Selector.

-> Selector
The right Selector.

-> Selector
A selector that is the combination of the left Selector and the right Selector through Preceded.

Combines two Selectors with the Preceded combinator.

Filters

data SelectorFilter

A type that sums up the different ways to filter a type selector: with an id (hash), a class, and an attribute.

Constructors

SHash Hash
A Hash object as filter.

SClass Class
A Class object as filter.

SAttrib Attrib
An Attrib object as filter.

SPseudo PseudoClass
A PseudoClass object as filter.

SNot Negation
A :not(…) clause that contains a simple selector to negate.

filters

Arguments

:: SelectorSequence
The given SelectorSequence to analyze.

-> [SelectorFilter]
The given list of SelectorFilters applied, this can be empty.

Obtain the list of filters that are applied in the given SelectorSequence.

filters'

Arguments

:: SelectorSequence
The given SelectorSequence to analyze.

-> [SelectorFilter]
The given list of SelectorFilters applied in reversed order, this can be empty.

Obtain the list of filters that are applied in the given SelectorSequence in reversed order.

addFilters

Arguments

:: SelectorSequence
The SelectorSequence to apply the filter on.

-> [SelectorFilter]
The list of SelectorFilters to apply on the SelectorSequence.

-> SelectorSequence
A modified SelectorSequence where we applied the list of SelectorFilters.

Add a given list of SelectorFilters to the given SelectorSequence. The filters are applied left-to-right.

(.@)

Arguments

:: SelectorSequence
The SelectorSequence to apply the filter on.

-> [SelectorFilter]
The list of SelectorFilters to apply on the SelectorSequence.

-> SelectorSequence
A modified SelectorSequence where we applied the list of SelectorFilters.

An infix variant of the addFilters function.

Namespaces

data Namespace

The namespace of a css selector tag. The namespace can be NAny (all possible namespaces), or a namespace with a given text (this text can be empty).

Constructors

NAny
A typeselector part that specifies that we accept all namespaces, in css denoted with *.

Namespace Text
A typselector part that specifies that we accept a certain namespace name.

pattern NEmpty :: Namespace

The empty namespace. This is not the wildcard namespace (*). This is a bidirectional namespace and can thus be used in expressions as well.

Type selectors

data ElementName

The element name of a css selector tag. The element name can be EAny (all possible tag names), or an element name with a given text.

Constructors

EAny
A typeselector part that specifies that we accept all element names, in css denoted with *.

ElementName Text
A typeselector part that specifies that we accept a certain element name.

data TypeSelector

A typeselector is a combination of a selector for a namespace, and a selector for an element name. One, or both can be a wildcard.

Constructors

TypeSelector

Fields

selectorNamespace :: Namespace

The selector for the namespace.

elementName :: ElementName

The selector for the element name.

pattern Universal :: TypeSelector

The universal type selector: a selector that matches all types in all namespaces (including the empty namespace). This pattern is bidirectional and thus can be used in expressions as well.

(.|)

Arguments

:: Namespace
The Namespace for the TypeSelector.

-> ElementName
The ElementName for the TypeSelector.

-> TypeSelector
A TypeSelector object constructed with the Namespace and ElementName.

Construct a TypeSelector with a given Namespace and ElementName.

Attributes

data Attrib

A css attribute can come in two flavors: either a constraint that the attribute should exists, or a constraint that a certain attribute should have a certain value (prefix, suffix, etc.).

Constructors

Exist AttributeName
A constraint that the given AttributeName should exist.

Attrib AttributeName AttributeCombinator AttributeValue
A constraint about the value associated with the given AttributeName.

data AttributeCombinator

The possible ways to match an attribute with a given value in a css selector.

Constructors

Exact
The attribute has exactly the value of the value, denoted with = in css.

Include
The attribute has a whitespace separated list of items, one of these items is the value, denoted with ~= in css.

DashMatch
The attribute has a hyphen separated list of items, the first item is the value, denoted with |= in css.

PrefixMatch
The value is a prefix of the value in the attribute, denoted with ^= in css.

SuffixMatch
The value is a suffix of the value in the attribute, denoted with $= in css.

SubstringMatch
The value is a substring of the value in the attribute, denoted with *= in css.

data AttributeName

An attribute name is a name that optionally has a namespace, and the name of the attribute.

Constructors

AttributeName

Fields

attributeNamespace :: Namespace

The namespace to which the attribute name belongs. This can be NAny as well.

attributeName :: Text

The name of the attribute over which we make a claim.

type AttributeValue = Text

We use Text as the type to store an attribute value.

(.=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted to be exactly the given value.

(.~=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute is a whitespace seperated list of items, and the value is one of these items.

(.|=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute is a dash seperated list of items, and the value is the first of these items.

(.^=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute has as prefix the given AttributeValue.

(.$=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute has as suffix the given AttributeValue.

(.*=)

Arguments

:: AttributeName
The name of the attribute to constraint.

-> AttributeValue
The value that constraints the attribute.

-> Attrib
The Attrib object we construct with the given name and value.

Create an Attrib where the given AttributeName is constrainted such that the attribute has as substring the given AttributeValue.

attrib

Arguments

:: AttributeCombinator
The AttributeCombinator that specifies the required relation between the attribute and a value.

-> AttributeName
The name of an attribute to filter.

-> AttributeValue
The value of the attribute to filter.

-> Attrib
The result is an Attrib object that will filter the given AttributeName with the given AttributeCombinator.

A flipped version of the Attrib data constructor, where one first specifies the conbinator, then the AttributeName and finally the value.

attributeCombinatorText

Arguments

:: AttributeCombinator
The AttributeCombinator for which we obtain the corresponding css selector text.

-> AttributeValue
The css selector text for the given AttributeCombinator.

Convert the given AttributeCombinator to its css selector counterpart.

Classes

newtype Class

A css class, this is wrapped in a data type. The type only wraps the class name, not the dot prefix.

Constructors

Class

Fields

unClass :: Text

Obtain the name from the class.

(...)

Arguments

:: SelectorSequence
The given 'SelectorSequence to filter.

-> Class
The given Class to filter the SelectorSequence further.

-> SelectorSequence
A SelectorSequence that is filtered additionally with the given Class.

Filter a given SelectorSequence with a given Class.

Hashes

newtype Hash

A css hash (used to match an element with a given id). The type only wraps the hash name, not the hash (#) prefix.

Constructors

Hash

Fields

unHash :: Text

Obtain the name from the hash.

(.#)

Arguments

:: SelectorSequence
The given SelectorSequence to filter.

-> Hash
The given Hash to filter the SelectorSequence further.

-> SelectorSequence
A SelectorSequence that is filtered additionally with the given Hash.

Filter a given SelectorSequence with a given Hash.

Negation

data Negation

A data type that contains all possible items that can be used in a :not(…) clause. Since a :not(…) cannot be nested in another :not(…), we see an SNot as a special case, and not as a PseudoClass.

Constructors

NTypeSelector TypeSelector
A TypeSelector for the :not(…) clause.

NHash Hash
A Hash for the :not(…) clause.

NClass Class
A Class for the :not(…) clause.

NAttrib Attrib
An Attrib for the :not(…) clause.

NPseudo PseudoClass
A PseudoClass for the :not(…) clause.

NPseudoElement PseudoElement
A PseudoElement for the :not(…) clause.

Nth items

data Nth

A data type that is used to select children and elements of type with the :nth-child, :nth-last-child, :nth-last-of-type and :nth-of-type. if the One is used as argument, then the pseudo classes are :first-child, :first-of-type, :last-child, and :last-of-type.

Constructors

Nth

Fields

linear :: Int

The linear part of the Nth object: the integral number before the n.

constant :: Int

The constant part of the Nth object.

pattern Even :: Nth

A pattern synonym that is used in CSS to specify a sequence that starts with two and each time increases with two.

pattern Odd :: Nth

A pattern synonym that is used in CSS to specify a sequence that starts with one and each time increases with two.

pattern One :: Nth

An Nth item that spans a collection with only 1 as value. This is used to transform :nth-child to :first-child for example.

nthValues

Arguments

:: Nth

-> [Int]
A list of one-based indexes that contain the items selected by the Nth object. The list can be infinite.

Obtain the one-based indices that match the given Nth object. The CSS3 selectors are one-based: the first child has index 1.

nthIsEmpty

Arguments

:: Nth
The given Nth object object to check.

-> Bool
True if the given Nth object does not contain any items; False otherwise.

Check if the given Nth object contains no items.

nthValues0

Arguments

:: Nth

-> [Int]
A list of zero-based indexes that contain the items selected by the Nth object. The list can be infinite.

Obtain the zero-based indices that match the given Nth object. One can use this for list/vector processing since the CSS3 selectors start with index 1. The nthValues1 can be used for one-based indexes.

nthValues1

Arguments

:: Nth

-> [Int]
A list of zero-based indexes that contain the items selected by the Nth object. The list can be infinite.

Obtain the one-based indices that match the given Nth object. The CSS3 selectors are one-based: the first child has index 1. This is an alias of the nthValues function.

normalizeNth

Arguments

:: Nth
The given Nth item to normalize.

-> Nth
The normalized variant of the given Nth object.

Normalize the given Nth object to a normalized one. If and only if the normalized variants are the same of two Nth objects, then these will produce the same list of values. Normalization is idempotent: calling normalizeNth on a normalized Nth will produce the same Nth.

nthContainsValue

Arguments

:: Nth
The given Nth object that specifies a sequence.

-> Int
The given index for which we check if it is contained in the given Nth object.

-> Bool
This function returns True if the given item is a member of the given Nth sequence; False otherwise.

Check if the given Nth object contains a given value.

Specificity

data SelectorSpecificity

A datastructure that specifies the selectivity of a css selector. The specificity is calculated based on three integers: a, b and c.

The specificity is calculated with 100*a+10*b+c where a, b and c count certain elements of the css selector.

Constructors

SelectorSpecificity Int Int Int
Create a SelectorSpecificity object with a given value for a, b, and c.

specificity

Arguments

:: ToCssSelector a

=> a
The object for which we evaluate the specificity.

-> Int
The specificity level as an Int value.

Calculate the specificity of a ToCssSelector type object. This is done by calculating the SelectorSpecificity object, and then calculating the value of that object.

specificityValue

Arguments

:: SelectorSpecificity
The SelectorSpecificity to calculate the specificity value from.

-> Int
The specificity level of the SelectorSpecificity. If the value is higher, the rules in the css selector take precedence.

Calculate the specificity value of the SelectorSpecificity

Read and write binary content

encode :: Binary a => a -> ByteString

Encode a value using binary serialisation to a lazy ByteString.

decode :: Binary a => ByteString -> a

Decode a value from a lazy ByteString, reconstructing the original structure.

compressEncode

Arguments

:: (Binary a, ToCssSelector a)

=> a
The object to turn into a compressed ByteString.

-> ByteString
A compressed binary representation of the given object.

Convert the given item to a compressed ByteString. This can be used to write to and read from a file for example. The econding format is not an official format: it is constructed based on the structure of the Haskell types. That stream is then passed through a gzip implementation.

compressEncodeWith

Arguments

:: (Binary a, ToCssSelector a)

=> CompressParams
The parameters that determine how to compress the ByteString.

-> a
The object to turn into a compressed ByteString.

-> ByteString
A compressed binary representation of the given object.

Convert the given item to a compressed ByteString. This can be used to write to and read from a file for example. The econding format is not an official format: it is constructed based on the structure of the Haskell types. That stream is then passed through a gzip implementation.

decompressDecode

Arguments

:: (Binary a, ToCssSelector a)

=> ByteString
A compressed binary representation of a ToCssSelector type.

-> a
The corresponding decompressed and decoded logic.

Convert the given item to a compressed ByteString. This can be used to write to and read from a file for example. The econding format is not an official format: it is constructed based on the structure of the Haskell types. That stream is then passed through a gzip implementation.