Css3.Selector.QuasiQuoters

Description

A module that defines a quasiquoter to parse a string to a css selector.

Documentation

csssel :: QuasiQuoter

A quasiquoter that can be used to construct a SelectorGroup for the given css selector. In case the css selector is invalid. A compiler error will be thrown (at compile time).

cssselFile :: QuasiQuoter

A quasiquoter that takes the content from the file, and then runs the content of that file as a csssel quasiquote.

parseCss

Arguments

:: String
The string to be parsed to a SelectorGroup

-> SelectorGroup
The selectorgroup that is the equivalent of the given String.

Parse the string to a SelectorGroup.